requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
pydub>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0

# Jupyter and Notebooks
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# Model Tracking and Logging
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.0

# Audio Augmentation
torch-audiomentations>=0.11.0
audiomentations>=0.30.0
